refactor: repository, service 계층에서 구현체가 하나 뿐인 인터페이스는 삭제
