+++ src/main/java/gift/controller/ProductApiController.java
@@ -3,7 +3,6 @@ package gift.controller;
 import gift.dto.ProductRequest;
 import gift.dto.ProductResponse;
 import gift.service.ProductService;
-import gift.service.ProductServiceImpl;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -22,7 +21,7 @@ public class ProductApiController {
 
     private final ProductService productService;
 
-    public ProductApiController(ProductServiceImpl productService) {
+    public ProductApiController(ProductService productService) {
         this.productService = productService;
     }
