--- src/main/java/gift/service/ProductServiceImpl.java
@@ -1,74 +1,0 @@
-package gift.service;
-
-import gift.domain.Product;
-import gift.dto.ProductMapper;
-import gift.dto.ProductRequest;
-import gift.dto.ProductResponse;
-import gift.exception.ErrorCode;
-import gift.exception.InvalidProductException;
-import gift.repository.ProductRepository;
-import java.util.List;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ProductServiceImpl implements ProductService {
-
-    private final ProductRepository productRepository;
-
-    public ProductServiceImpl(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    @Override
-    public ProductResponse getProductById(Long productId) {
-        Product product = productRepository.findById(productId)
-            .orElseThrow(() -> new InvalidProductException(ErrorCode.NOT_EXISTS_PRODUCT));
-
-        return ProductResponse.of(product.getName(), product.getPrice(), product.getImageUrl());
-    }
-
-    @Override
-    public Long save(ProductRequest request) {
-        validateRequest(request);
-
-        Long id = productRepository.save(ProductMapper.toEntity(request));
-        return id;
-    }
-
-    @Override
-    public void update(ProductRequest request) {
-        if(request==null || request.id()==null)
-            throw new InvalidProductException(ErrorCode.INVALID_PRODUCT_UPDATE_REQUEST);
-
-        productRepository.findById(request.id())
-            .orElseThrow(() -> new InvalidProductException(ErrorCode.NOT_EXISTS_PRODUCT));
-
-        productRepository.update(request);
-    }
-
-    @Override
-    public void deleteById(Long productId) {
-        productRepository.deleteById(productId);
-    }
-
-    @Override
-    public List<ProductResponse> getProductList() {
-        return productRepository.findAll().stream()
-            .map(ProductMapper::toResponse)
-            .toList();
-    }
-
-    private void validateRequest(ProductRequest request) {
-        if (request == null) {
-            throw new InvalidProductException(ErrorCode.NOT_EXISTS_PRODUCT);
-        }
-
-        if (request.name() == null || request.imageUrl() == null) {
-            throw new InvalidProductException(ErrorCode.INVALID_PRODUCT_REQUEST);
-        }
-
-        if (request.price() < 0) {
-            throw new InvalidProductException(ErrorCode.INVALID_PRODUCT_PRICE);
-        }
-    }
-}
